requirements.txt
alabaster==0.7.12
amqp==5.0.6
anytree==2.8.0
asgiref==3.4.1
astroid==2.2.5
attrs==20.3.0
autopep8==1.4.4
Babel==2.9.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.2.1
bullet==2.2.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2019.11.28
cffi==1.14.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cliff==3.5.0
cmd2==1.4.0
colorama==0.4.4
configobj==5.0.6
configparser==5.0.1
crispy-bootstrap5==0.3
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.7
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==21.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-import-export==2.5.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-paypal==1.1.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-taggit==1.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dnspython==2.0.0
docutils==0.16
et-xmlfile==1.1.0
filelock==3.0.12
gnupg==2.3.1
gunicorn==20.0.4
html5lib==1.1
httpie==2.4.0
humanfriendly==9.1
idna==2.10
imagesize==1.2.0
importlib-metadata==2.1.2
ipwhois==1.2.0
isort==4.3.20
Jinja2==2.11.2
keyring==21.8.0
kombu==5.1.0
lazy-object-proxy==1.4.1
lxml==4.6.2
Markdown==3.1.1
markdown2==2.3.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.8
pbr==5.5.1
pexpect==4.8.0
pi==0.1.2
Pillow==8.3.2
Pillow-PIL==0.1.dev0
pipenv==2018.11.26
pkginfo==1.6.1
pluggy==0.13.1
prettytable==0.7.2
prompt-toolkit==3.0.19
property-manager==3.0
psutil==5.8.0
psycopg2==2.8.4
psycopg2-binary==2.8.5
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.2.0
pygame==2.1.2
Pygments==2.7.3
pylint==2.3.1
pyparsing==2.4.7
pyperclip==1.8.1
Pyphen==0.10.0
pyreadline==2.1
PySocks==1.7.1
python-secrets==20.11.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.1
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.2
stevedore==3.3.0
tablib==3.0.0
tinycss2==1.1.0
toml==0.10.2
tox==3.21.0
tqdm==4.55.2
twine==3.3.0
typed-ast==1.5.1
typing-extensions==3.7.4.3
urllib3==1.26.2
verboselogs==1.7
vine==5.0.0
virtualenv==16.4.3
virtualenv-clone==0.5.3
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.11.1
xkcdpass==1.17.6
xlrd==2.0.1
xlwt==1.3.0
yamlreader==3.0.4
zipp==3.4.0
